Fix circular dependency introduced by tests for AXISC-1627
The tests for AXISC-1627 created a circular link dependency between
axiom and guththila.  Since the tests provided with the fix relied on
axiom calls, it made sense to move the tests to the axiom subproject.
This has the benefit of testing other parsers for this issue, instead of only
guththila.